+++ src/frigate_event_processor/app_configuration.py
